+++ LvqEmn/LvqLib/LvqDataset.cpp
@@ -284,7 +284,7 @@ std::vector<int> LvqDataset::GetTestSubset(int fold, int foldcount) const {
 	int foldStart = fold * pointCount / foldcount;
 	int foldEnd = (fold+1) * pointCount / foldcount;
 	std::vector<int> retval(foldEnd-foldStart);
-	for(int i=0;i<retval.size();++i)
+	for(size_t i=0;i<retval.size();++i)
 		retval[i] = foldStart + i;
 	return retval;
 }
